+++ src/integrate/scvi/script.py
@@ -61,15 +61,17 @@ def main():
 
     if par['var_input']:
         # Subset to HVG
-        adata = subset_vars(adata, subset_col=par["var_input"])
+        adata_subset = subset_vars(adata, subset_col=par["var_input"]).copy()
+    else:
+        adata_subset = adata.copy()
 
     check_validity_anndata(
-        adata, par['input_layer'], par['obs_batch'],
+        adata_subset, par['input_layer'], par['obs_batch'],
         par["n_obs_min_count"], par["n_var_min_count"]
-        )
+    )
     # Set up the data
     scvi.model.SCVI.setup_anndata(
-        adata,
+        adata_subset,
         batch_key=par['obs_batch'],
         layer=par['input_layer'],
         labels_key=par['obs_labels'],
@@ -80,7 +82,7 @@ def main():
 
     # Set up the model
     vae_uns = scvi.model.SCVI(
-        adata,
+        adata_subset,
         n_hidden=par["n_hidden_nodes"],
         n_latent=par["n_dimensions_latent_space"],
         n_layers=par["n_hidden_layers"],
@@ -103,7 +105,7 @@ def main():
 
     # Train the model
     vae_uns.train(
-        max_epochs = par['max_epochs'],
+        max_epochs=par['max_epochs'],
         early_stopping=par['early_stopping'],
         early_stopping_monitor=par['early_stopping_monitor'],
         early_stopping_patience=par['early_stopping_patience'],
@@ -112,7 +114,7 @@ def main():
         check_val_every_n_epoch=1,
         accelerator="auto",
     )
-    #Note: train_size=1.0 should give better results, but then can't do early_stopping on validation set
+    # Note: train_size=1.0 should give better results, but then can't do early_stopping on validation set
 
     # Get the latent output
     adata.obsm[par['obsm_output']] = vae_uns.get_latent_representation()